async_python/pyproject.toml
[tool.poetry]
name = "async_python"
version = "0.1.0"
description = ""
authors = ["Steve Dignam <steve@dignam.xyz>"]

[tool.poetry.dependencies]
python = "^3.10"
starlette = "^0.20.4"
uvicorn = "^0.18.3"
asyncpg = "^0.26.0"
orjson = "^3.8.0"
gunicorn = {extras = ["gthread"], version = "^20.1.0"}
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
asyncpg-stubs = "^0.26.4"
mypy = "^0.982"

[tool.poetry.dev-dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
